Add advisory vulnerability assertions to the REST API. Refactor where/how to fetch assertions from AdvisoryVulnerablity instead of the Advisory level.
