+++ common/src/advisory/mod.rs
@@ -1,24 +1,23 @@
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
+use utoipa::ToSchema;
 
-#[derive(Debug, Serialize, Deserialize, Clone)]
+#[derive(Debug, Serialize, Deserialize, Clone, ToSchema)]
 pub struct AdvisoryVulnerabilityAssertions {
     pub assertions: HashMap<String, Vec<Assertion>>,
 }
 
-#[derive(Serialize, Deserialize, Clone, Debug)]
+#[derive(Serialize, Deserialize, Clone, Debug, ToSchema)]
+#[serde(rename_all = "lowercase")]
 pub enum Assertion {
     Affected {
-        vulnerability: String,
         start_version: String,
         end_version: String,
     },
     NotAffected {
-        vulnerability: String,
         version: String,
     },
     Fixed {
-        vulnerability: String,
         version: String,
     },
 }
