+++ modules/graph/src/endpoints/mod.rs
@@ -52,7 +52,9 @@ pub fn configure(config: &mut web::ServiceConfig) {
     ),
     components(schemas(
         crate::model::advisory::AdvisoryDetails,
-        crate::model::advisory::AdvisoryVulnerabilitySummary,
+        crate::model::advisory::AdvisoryVulnerability,
+        trustify_common::advisory::AdvisoryVulnerabilityAssertions,
+        trustify_common::advisory::Assertion,
     )),
     tags()
 )]
