+++ modules/graph/src/graph/advisory/advisory_vulnerability.rs
@@ -4,7 +4,13 @@ use crate::graph::advisory::not_affected_package_version::NotAffectedPackageVers
 use crate::graph::advisory::AdvisoryContext;
 use crate::graph::error::Error;
 use crate::graph::vulnerability::VulnerabilityContext;
-use sea_orm::{ActiveModelTrait, ColumnTrait, EntityTrait, QueryFilter, Set};
+use sea_orm::{
+    ActiveModelTrait, ColumnTrait, EntityTrait, FromQueryResult, QueryFilter, QuerySelect,
+    RelationTrait, Set,
+};
+use sea_query::JoinType;
+use std::collections::HashMap;
+use trustify_common::advisory::{AdvisoryVulnerabilityAssertions, Assertion};
 use trustify_common::db::Transactional;
 use trustify_common::purl::Purl;
 use trustify_cvss::cvss3::Cvss3Base;
@@ -267,5 +273,375 @@ impl<'g> AdvisoryVulnerabilityContext<'g> {
         };
 
         Ok(model.insert(&self.advisory.graph.db).await?.into())
+    }
+
+    pub async fn vulnerability_assertions<TX: AsRef<Transactional>>(
+        &self,
+        tx: TX,
+    ) -> Result<AdvisoryVulnerabilityAssertions, Error> {
+        let affected = self.affected_assertions(&tx).await?;
+        let not_affected = self.not_affected_assertions(&tx).await?;
+        let fixed = self.fixed_assertions(&tx).await?;
+
+        let mut merged = affected.assertions.clone();
+
+        for (package_key, assertions) in not_affected.assertions {
+            merged
+                .entry(package_key)
+                .or_insert(Vec::default())
+                .extend_from_slice(&assertions)
+        }
+
+        for (package_key, assertions) in fixed.assertions {
+            merged
+                .entry(package_key)
+                .or_insert(Vec::default())
+                .extend_from_slice(&assertions)
+        }
+
+        Ok(AdvisoryVulnerabilityAssertions { assertions: merged })
+    }
+
+    pub async fn affected_assertions<TX: AsRef<Transactional>>(
+        &self,
+        tx: TX,
+    ) -> Result<AdvisoryVulnerabilityAssertions, Error> {
+        #[derive(FromQueryResult, Debug)]
+        struct AffectedVersion {
+            ty: String,
+            namespace: Option<String>,
+            name: String,
+            start: String,
+            end: String,
+            vulnerability: String,
+            identifier: String,
+            location: String,
+            sha256: String,
+        }
+
+        let mut affected_version_ranges = entity::affected_package_version_range::Entity::find()
+            .column_as(entity::package::Column::Type, "ty")
+            .column_as(entity::package::Column::Namespace, "namespace")
+            .column_as(entity::package::Column::Name, "name")
+            .column_as(entity::package_version_range::Column::Start, "start")
+            .column_as(entity::package_version_range::Column::End, "end")
+            .column_as(entity::vulnerability::Column::Identifier, "vulnerability")
+            .column_as(entity::advisory::Column::Identifier, "identifier")
+            .column_as(entity::advisory::Column::Location, "location")
+            .column_as(entity::advisory::Column::Sha256, "sha256")
+            .join(
+                JoinType::Join,
+                entity::affected_package_version_range::Relation::PackageVersionRange.def(),
+            )
+            .join(
+                JoinType::Join,
+                entity::affected_package_version_range::Relation::Advisory.def(),
+            )
+            .join(
+                JoinType::Join,
+                entity::package_version_range::Relation::Package.def(),
+            )
+            .join(
+                JoinType::Join,
+                entity::advisory_vulnerability::Relation::Advisory
+                    .def()
+                    .rev(),
+            )
+            .join(
+                JoinType::Join,
+                entity::advisory_vulnerability::Relation::Vulnerability.def(),
+            )
+            .filter(
+                entity::affected_package_version_range::Column::AdvisoryId
+                    .eq(self.advisory.advisory.id),
+            )
+            .into_model::<AffectedVersion>()
+            .all(&self.advisory.graph.connection(&tx))
+            .await?;
+
+        let mut assertions = HashMap::new();
+
+        for each in affected_version_ranges {
+            let package_key = Purl {
+                ty: each.ty,
+                namespace: each.namespace,
+                name: each.name,
+                version: None,
+                qualifiers: Default::default(),
+            }
+            .to_string();
+
+            let mut package_assertions = assertions.entry(package_key.clone()).or_insert(vec![]);
+
+            package_assertions.push(Assertion::Affected {
+                start_version: each.start,
+                end_version: each.end,
+            });
+        }
+
+        Ok(AdvisoryVulnerabilityAssertions { assertions })
+    }
+
+    pub async fn not_affected_assertions<TX: AsRef<Transactional>>(
+        &self,
+        tx: TX,
+    ) -> Result<AdvisoryVulnerabilityAssertions, Error> {
+        #[derive(FromQueryResult, Debug)]
+        struct NotAffectedVersion {
+            ty: String,
+            namespace: Option<String>,
+            name: String,
+            version: String,
+            vulnerability: String,
+            identifier: String,
+            location: String,
+            sha256: String,
+        }
+
+        let mut not_affected_versions = entity::not_affected_package_version::Entity::find()
+            .column_as(entity::package::Column::Type, "ty")
+            .column_as(entity::package::Column::Namespace, "namespace")
+            .column_as(entity::package::Column::Name, "name")
+            .column_as(entity::package_version::Column::Version, "version")
+            .column_as(entity::vulnerability::Column::Identifier, "vulnerability")
+            .column_as(entity::advisory::Column::Identifier, "identifier")
+            .column_as(entity::advisory::Column::Location, "location")
+            .column_as(entity::advisory::Column::Sha256, "sha256")
+            .join(
+                JoinType::Join,
+                entity::not_affected_package_version::Relation::PackageVersion.def(),
+            )
+            .join(
+                JoinType::Join,
+                entity::not_affected_package_version::Relation::Advisory.def(),
+            )
+            .join(
+                JoinType::Join,
+                entity::package_version::Relation::Package.def(),
+            )
+            .join(
+                JoinType::Join,
+                entity::advisory_vulnerability::Relation::Advisory
+                    .def()
+                    .rev(),
+            )
+            .join(
+                JoinType::Join,
+                entity::advisory_vulnerability::Relation::Vulnerability.def(),
+            )
+            .filter(
+                entity::not_affected_package_version::Column::AdvisoryId
+                    .eq(self.advisory.advisory.id),
+            )
+            .into_model::<NotAffectedVersion>()
+            .all(&self.advisory.graph.connection(&tx))
+            .await?;
+
+        let mut assertions = HashMap::new();
+
+        for each in not_affected_versions {
+            let package_key = Purl {
+                ty: each.ty,
+                namespace: each.namespace,
+                name: each.name,
+                version: None,
+                qualifiers: Default::default(),
+            }
+            .to_string();
+
+            let mut package_assertions = assertions.entry(package_key.clone()).or_insert(vec![]);
+
+            package_assertions.push(Assertion::NotAffected {
+                version: each.version,
+            });
+        }
+
+        Ok(AdvisoryVulnerabilityAssertions { assertions })
+    }
+
+    pub async fn fixed_assertions<TX: AsRef<Transactional>>(
+        &self,
+        tx: TX,
+    ) -> Result<AdvisoryVulnerabilityAssertions, Error> {
+        #[derive(FromQueryResult, Debug)]
+        struct FixedVersion {
+            ty: String,
+            namespace: Option<String>,
+            name: String,
+            version: String,
+            vulnerability: String,
+            identifier: String,
+            location: String,
+            sha256: String,
+        }
+
+        let mut fixed_versions = entity::fixed_package_version::Entity::find()
+            .column_as(entity::package::Column::Type, "ty")
+            .column_as(entity::package::Column::Namespace, "namespace")
+            .column_as(entity::package::Column::Name, "name")
+            .column_as(entity::package_version::Column::Version, "version")
+            .column_as(entity::vulnerability::Column::Identifier, "vulnerability")
+            .column_as(entity::advisory::Column::Identifier, "identifier")
+            .column_as(entity::advisory::Column::Location, "location")
+            .column_as(entity::advisory::Column::Sha256, "sha256")
+            .join(
+                JoinType::Join,
+                entity::fixed_package_version::Relation::PackageVersion.def(),
+            )
+            .join(
+                JoinType::Join,
+                entity::fixed_package_version::Relation::Advisory.def(),
+            )
+            .join(
+                JoinType::Join,
+                entity::package_version::Relation::Package.def(),
+            )
+            .join(
+                JoinType::Join,
+                entity::advisory_vulnerability::Relation::Advisory
+                    .def()
+                    .rev(),
+            )
+            .join(
+                JoinType::Join,
+                entity::advisory_vulnerability::Relation::Vulnerability.def(),
+            )
+            .filter(entity::fixed_package_version::Column::AdvisoryId.eq(self.advisory.advisory.id))
+            .into_model::<FixedVersion>()
+            .all(&self.advisory.graph.connection(&tx))
+            .await?;
+
+        let mut assertions = HashMap::new();
+
+        for each in fixed_versions {
+            let package_key = Purl {
+                ty: each.ty,
+                namespace: each.namespace,
+                name: each.name,
+                version: None,
+                qualifiers: Default::default(),
+            }
+            .to_string();
+
+            let mut package_assertions = assertions.entry(package_key.clone()).or_insert(vec![]);
+
+            package_assertions.push(Assertion::Fixed {
+                version: each.version,
+            });
+        }
+
+        Ok(AdvisoryVulnerabilityAssertions { assertions })
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use crate::graph::Graph;
+    use test_log::test;
+    use trustify_common::advisory::Assertion;
+    use trustify_common::db::{Database, Transactional};
+
+    #[test(tokio::test)]
+    async fn advisory_affected_vulnerability_assertions() -> Result<(), anyhow::Error> {
+        let db = Database::for_test("advisory_affected_vulnerability_assertions").await?;
+        let system = Graph::new(db);
+
+        let advisory = system
+            .ingest_advisory(
+                "RHSA-GHSA-1",
+                "http://db.com/rhsa-ghsa-2",
+                "2",
+                (),
+                Transactional::None,
+            )
+            .await?;
+
+        let advisory_vulnerability = advisory
+            .link_to_vulnerability("CVE-42", Transactional::None)
+            .await?;
+
+        advisory_vulnerability
+            .ingest_affected_package_range(
+                "pkg://maven/io.quarkus/quarkus-core".try_into()?,
+                "1.0.2",
+                "1.2.0",
+                Transactional::None,
+            )
+            .await?;
+
+        advisory_vulnerability
+            .ingest_not_affected_package_version(
+                "pkg://maven/.io.quarkus/quarkus-core@1.1.9".try_into()?,
+                Transactional::None,
+            )
+            .await?;
+
+        let affected = advisory_vulnerability
+            .affected_assertions(Transactional::None)
+            .await?;
+
+        assert_eq!(1, affected.assertions.len());
+
+        Ok(())
+    }
+
+    #[test(tokio::test)]
+    async fn advisory_not_affected_vulnerability_assertions() -> Result<(), anyhow::Error> {
+        let db = Database::for_test("advisory_not_affected_vulnerability_assertions").await?;
+        let system = Graph::new(db);
+
+        let advisory = system
+            .ingest_advisory(
+                "RHSA-GHSA-1",
+                "http://db.com/rhsa-ghsa-2",
+                "2",
+                (),
+                Transactional::None,
+            )
+            .await?;
+
+        let advisory_vulnerability = advisory
+            .link_to_vulnerability("INTERAL-77", Transactional::None)
+            .await?;
+
+        advisory_vulnerability
+            .ingest_affected_package_range(
+                "pkg://maven/io.quarkus/quarkus-core".try_into()?,
+                "1.0.2",
+                "1.2.0",
+                Transactional::None,
+            )
+            .await?;
+
+        advisory_vulnerability
+            .ingest_not_affected_package_version(
+                "pkg://maven/io.quarkus/quarkus-core@1.1.9".try_into()?,
+                Transactional::None,
+            )
+            .await?;
+
+        let not_affected = advisory_vulnerability
+            .not_affected_assertions(Transactional::None)
+            .await?;
+
+        assert_eq!(1, not_affected.assertions.len());
+
+        let pkg_assertions = not_affected
+            .assertions
+            .get(&"pkg://maven/io.quarkus/quarkus-core".to_string());
+
+        assert!(pkg_assertions.is_some());
+
+        let pkg_assertions = pkg_assertions.unwrap();
+
+        assert_eq!(1, pkg_assertions.len());
+
+        let assertion = &pkg_assertions[0];
+
+        assert!(matches!( assertion, Assertion::NotAffected {version, ..}
+            if version == "1.1.9"
+        ));
+
+        Ok(())
     }
 }
